% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{epi_qmd}
\alias{epi_qmd}
\title{Use quarto report or book templates}
\usage{
epi_qmd(
  path = ".",
  quiet = TRUE,
  project_type = "document",
  add_project = (project_type == "book")
)
}
\arguments{
\item{path}{The path to the quarto document}

\item{quiet}{Suppress messages}

\item{project_type}{A character of length 1, specifying which qmd repo to download. Could be one of ("document", "book"). Default is "document".}

\item{add_project}{Boolean. Add a Rstudio project to the new document folder.}
}
\description{
Download and use document/book quarto template.
}
\examples{
\dontrun{
epi_qmd(path = ".", quiet = TRUE, project_type = "document")
epi_qmd(path = ".", quiet = TRUE, project_type = "book")
}
}
