% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{epi_render}
\alias{epi_render}
\title{Render rmarkdown or quarto documents/}
\usage{
epi_render(
  path,
  output_format = "all",
  output_dir = NULL,
  apply_rules = TRUE,
  ...
)
}
\arguments{
\item{path}{Character vector, paths fo files to render}

\item{output_format}{Output format, default is all ie renders to all output}

\item{output_dir}{Where to save all output files. Default is the same folder as \code{path}}

\item{apply_rules}{Boolean. Whether to apply rules in _outputs.yml files or not. Default is TRUE.}

\item{...}{other parameters passed to \code{rmarkdown::render} or \code{quarto::render}}
}
\description{
Provide a wrapper to \code{rmarkdown::render} and \code{quarto::quarto_render}.
}
\section{Managing outputs with _outputs.yml}{
NA
}

\examples{
\dontrun{
epi_qmd()
epi_render(path = "template.qmd")
}
}
