% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usetemplates.R
\name{epids_qmd}
\alias{epids_qmd}
\title{Use quarto EpiDS templates}
\usage{
epids_qmd(
  dir = ".",
  quiet = FALSE,
  project_type = "document",
  add_project = (project_type == "book")
)
}
\arguments{
\item{dir}{The directory where to download the quarto document}

\item{quiet}{Suppress warnings and messages}

\item{project_type}{A character of length 1, specifying which qmd type to download. Could be one of ("document", "book"). Default is "document".}

\item{add_project}{Boolean. Add a Rstudio project to the document folder.}
}
\description{
Download and use the EpiDS document/book quarto template.
}
\details{
The provided directory should be empty.
}
\examples{
\dontrun{
epids_qmd_doc(dir = ".", quiet = TRUE)
epids_qmd_book(dir = ".", quiet = TRUE)
}
}
