% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitemplates-package.R
\docType{package}
\name{epitemplates-package}
\alias{epitemplates}
\alias{epitemplates-package}
\title{Generate templates for EpiDS documents}
\description{
A collection of Rmarkdown templates and R functions to use quarto
templates within Rstudio and VSCodes.
}
\author{
\strong{Maintainer}: Hugo Soubrier \email{hugo.soubrier@epicentre.msf.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Other contributors:
\itemize{
  \item Yves Amevoin \email{yves.amevoin@epicentre.msf.org} [contributor]
}

}
\keyword{internal}
